+++ Editor/Helpers/HeadlessSettings.cs
@@ -26,7 +26,7 @@ namespace BuildManager {
         [Serializable]
         public class GogHeadless {
             [Tooltip("Fill with Defines that should be forced enabled for Headless Builds. Other Defines will be disabled.")]
-            public string[] enabledDefinesOverwrite = new string[] { "GOGGALAXY", "DISABLESTEAMWORKS" };
+            public string[] enabledDefinesOverwrite = new string[] { "GOGGALAXY" };
         }
     }
 }