+++ CHANGELOG.md
@@ -14,3 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added use of a more generic way to create settings files in the user scope instead of the package scope
 - removed unneeded .asset
+
+## [0.1.2] - 2022-12-06
+### Added
+- added default defines for steam and gog
+- obsolete targetgroups are now filtered
+
+### Changed
+- remove DISABLESTEAMWORKS define from headless default defines
+
+### Fixed
+- fixed some UI issues when steam credentials & app config were not provided